// src/app/models/categories.enum.ts
// Categories to feed into the API search. Most, but not all, require cmtype=page.

export enum Categories {
    USPresidents = "Wikipedia_categories_named_after_Presidents_of_the_United_States",
    SecretSocieties = "Secret_societies&cmtype=page",
    Animals = "Set_indices_on_animal_common_names&cmtype=page",
    Wars = "Warfare_post-1945&cmtype=subcat",
    SingersSuicide = "Singers_who_committed_suicide&cmtype=page",
    AssRulers = "Assassinated_heads_of_state&cmtype=page",
    IntAgencies = "Intelligence_agencies&cmtype=page",
}